Move share endpoint, add tests
