+++ talentmap_api/common/tests/test_sharing.py
@@ -1,4 +1,5 @@
 import pytest
+import json
 
 from django.contrib.auth.models import User
 
@@ -11,6 +12,18 @@ from rest_framework import status
 def test_sharing_fixture():
     mommy.make('position.Position', id=2)
     mommy.make('auth.User', username="test", email="test@state.gov")
+
+
+@pytest.fixture
+def test_sharing_update_fixture(authorized_user):
+    mommy.make('position.Position', id=2)
+    sending_user = mommy.make('auth.User', username="test", email="test@state.gov")
+    mommy.make('user_profile.Sharable',
+               id=1,
+               sharing_user=sending_user.profile,
+               receiving_user=authorized_user.profile,
+               sharable_model="position.Position",
+               sharable_id=2)
 
 
 @pytest.mark.django_db()
@@ -27,11 +40,13 @@ def test_sharing_fixture():
     ({"mode": "internal"}, status.HTTP_400_BAD_REQUEST),
     ({"mode": "internal", "email": "a@a.c"}, status.HTTP_400_BAD_REQUEST),
     ({"mode": "internal", "id": 1, "type": "position"}, status.HTTP_400_BAD_REQUEST),
+    ({"mode": "internal", "email": "test@state.gov", "type": "position", "id": 1}, status.HTTP_404_NOT_FOUND),
+    ({"mode": "internal", "email": "dne@state.gov", "type": "position", "id": 2}, status.HTTP_404_NOT_FOUND),
     ({"mode": "internal", "email": "test@state.gov", "type": "position", "id": 2}, status.HTTP_202_ACCEPTED),
 ])
 @pytest.mark.usefixtures("test_sharing_fixture")
 def test_payload_validation(authorized_client, authorized_user, payload, resp):
-    response = authorized_client.post("/api/v1/share/", payload)
+    response = authorized_client.post("/api/v1/accounts/profile/share/", payload)
 
     assert response.status_code == resp
 
@@ -42,7 +57,7 @@ def test_internal_share(authorized_client, authorized_user):
     user_profile = User.objects.get(email="test@state.gov").profile
     assert user_profile.received_shares.count() == 0
 
-    response = authorized_client.post("/api/v1/share/", {
+    response = authorized_client.post("/api/v1/accounts/profile/share/", {
         "mode": "internal",
         "email": "test@state.gov",
         "type": "position",
@@ -52,3 +67,19 @@ def test_internal_share(authorized_client, authorized_user):
     assert response.status_code == status.HTTP_202_ACCEPTED
     user_profile.refresh_from_db()
     assert user_profile.received_shares.count() == 1
+
+
+@pytest.mark.django_db(transaction=True)
+@pytest.mark.usefixtures("test_sharing_update_fixture")
+def test_update_internal_share(authorized_client, authorized_user):
+    assert authorized_user.profile.received_shares.count() == 1
+    share = authorized_user.profile.received_shares.first()
+    assert not share.read
+
+    response = authorized_client.patch(f"/api/v1/accounts/profile/share/1/", data=json.dumps({
+        "read": True
+    }), content_type="application/json")
+
+    assert response.status_code == status.HTTP_200_OK
+    share.refresh_from_db()
+    assert share.read
